backend/requirements.txt
fastapi
uvicorn
openai
python-dotenv
redis
boto3
tenacity
pyjwt
psycopg2-binary
sqlalchemy
pytest
pytest-asyncio
httpx
requests
